backend/root/requirements.txt
# backend/root/requirements.txt

annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coverage==7.4.4
cryptography==44.0.2
Django==5.0.4
django-allauth==65.4.1
django-appconf==1.1.0
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-extensions==3.2.3
django_filter==25.1
django-select2==8.3.0
django-user-management==18.0.0
djangorestframework==3.10.3
idna==3.10
incuna-mail==4.1.1
incuna-pigeon==0.1.0
mysql-connector-python==9.2.0
mysqlclient==2.2.1
pillow==11.1.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.3.0
